ml/requirements.txt
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
imblearn==0.0
imbalanced-learn==0.10.1
matplotlib==3.7.1
seaborn==0.12.2
joblib==1.2.0
requests==2.28.2
beautifulsoup4==4.11.2
python-whois==0.7.3
dnspython==2.3.0
tqdm==4.65.0
flask==2.2.3
flask-cors==3.0.10
